test(patina): split preset tests into a sibling module
